added anonymous to do-not-delete list

+++ security/securityReplication/test/artifactory/SecurityTestApi.groovy
@@ -83,7 +83,7 @@ class SecurityTestApi {
     static def deleteAllUsers () {
         Collection<String> userNames = art.security().userNames()
         for (String userName : userNames) {
-            if (userName != "admin" && userName != "access-admin") {
+            if (userName != "anonymous" && userName != "admin" && userName != "access-admin") {
                 art.security().deleteUser(userName)
             }
         }
